pyproject.toml
[tool.poetry]
name = "poc-python-socketio-ns"
version = "0.1.0"
description = ""
authors = ["Mech <mechsix@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-socketio = "4.3.1"
tornado = "6.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
